Move binary image descriptors
