+++ sycl/include/sycl/detail/pi.h
@@ -22,6 +22,13 @@
 #ifndef _PI_H_
 #define _PI_H_
 
+// Ideally, we should only include compiler.h where it is actually used, but
+// since PI interfaces depend on some compiler interfaces we have to include
+// them here.
+// Once we have moved to UR where UR interfaces are not dependent on compiler
+// interfaces, we can clean this up and update our includes.
+#include <sycl/detail/compiler.h>
+
 // Every single change in PI API should be accompanied with the minor
 // version increase (+1). In the cases where backward compatibility is not
 // maintained there should be a (+1) change to the major version in
@@ -976,123 +983,6 @@ using pi_device_exec_capabilities = pi_bitfield;
 static constexpr pi_device_exec_capabilities PI_EXEC_KERNEL = (1 << 0);
 static constexpr pi_device_exec_capabilities PI_EXEC_NATIVE_KERNEL = (1 << 1);
 
-// Entry type, matches OpenMP for compatibility
-struct _pi_offload_entry_struct {
-  void *addr;
-  char *name;
-  size_t size;
-  int32_t flags;
-  int32_t reserved;
-};
-
-using _pi_offload_entry = _pi_offload_entry_struct *;
-
-// A type of a binary image property.
-typedef enum {
-  PI_PROPERTY_TYPE_UNKNOWN,
-  PI_PROPERTY_TYPE_UINT32,     // 32-bit integer
-  PI_PROPERTY_TYPE_BYTE_ARRAY, // byte array
-  PI_PROPERTY_TYPE_STRING      // null-terminated string
-} pi_property_type;
-
-// Device binary image property.
-// If the type size of the property value is fixed and is no greater than
-// 64 bits, then ValAddr is 0 and the value is stored in the ValSize field.
-// Example - PI_PROPERTY_TYPE_UINT32, which is 32-bit
-struct _pi_device_binary_property_struct {
-  char *Name;       // null-terminated property name
-  void *ValAddr;    // address of property value
-  uint32_t Type;    // _pi_property_type
-  uint64_t ValSize; // size of property value in bytes
-};
-
-typedef _pi_device_binary_property_struct *pi_device_binary_property;
-
-// Named array of properties.
-struct _pi_device_binary_property_set_struct {
-  char *Name;                                // the name
-  pi_device_binary_property PropertiesBegin; // array start
-  pi_device_binary_property PropertiesEnd;   // array end
-};
-
-typedef _pi_device_binary_property_set_struct *pi_device_binary_property_set;
-
-/// Types of device binary.
-using pi_device_binary_type = uint8_t;
-// format is not determined
-static constexpr pi_device_binary_type PI_DEVICE_BINARY_TYPE_NONE = 0;
-// specific to a device
-static constexpr pi_device_binary_type PI_DEVICE_BINARY_TYPE_NATIVE = 1;
-// portable binary types go next
-// SPIR-V
-static constexpr pi_device_binary_type PI_DEVICE_BINARY_TYPE_SPIRV = 2;
-// LLVM bitcode
-static constexpr pi_device_binary_type PI_DEVICE_BINARY_TYPE_LLVMIR_BITCODE = 3;
-
-// Device binary descriptor version supported by this library.
-static const uint16_t PI_DEVICE_BINARY_VERSION = 1;
-
-// The kind of offload model the binary employs; must be 4 for SYCL
-static const uint8_t PI_DEVICE_BINARY_OFFLOAD_KIND_SYCL = 4;
-
-/// Extension to denote native support of assert feature by an arbitrary device
-/// piDeviceGetInfo call should return this extension when the device supports
-/// native asserts if supported extensions' names are requested
-#define PI_DEVICE_INFO_EXTENSION_DEVICELIB_ASSERT "cl_intel_devicelib_assert"
-
-/// This struct is a record of the device binary information. If the Kind field
-/// denotes a portable binary type (SPIR-V or LLVM IR), the DeviceTargetSpec
-/// field can still be specific and denote e.g. FPGA target. It must match the
-/// __tgt_device_image structure generated by the clang-offload-wrapper tool
-/// when their Version field match.
-struct pi_device_binary_struct {
-  /// version of this structure - for backward compatibility;
-  /// all modifications which change order/type/offsets of existing fields
-  /// should increment the version.
-  uint16_t Version;
-  /// the type of offload model the binary employs; must be 4 for SYCL
-  uint8_t Kind;
-  /// format of the binary data - SPIR-V, LLVM IR bitcode,...
-  uint8_t Format;
-  /// null-terminated string representation of the device's target architecture
-  /// which holds one of:
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_UNKNOWN - unknown
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV32 - general value for 32-bit OpenCL
-  /// devices
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV64 - general value for 64-bit OpenCL
-  /// devices
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV64_X86_64 - 64-bit OpenCL CPU device
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV64_GEN - GEN GPU device (64-bit
-  /// OpenCL)
-  /// __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV64_FPGA - 64-bit OpenCL FPGA device
-  const char *DeviceTargetSpec;
-  /// a null-terminated string; target- and compiler-specific options
-  /// which are suggested to use to "compile" program at runtime
-  const char *CompileOptions;
-  /// a null-terminated string; target- and compiler-specific options
-  /// which are suggested to use to "link" program at runtime
-  const char *LinkOptions;
-  /// Pointer to the manifest data start
-  const char *ManifestStart;
-  /// Pointer to the manifest data end
-  const char *ManifestEnd;
-  /// Pointer to the target code start
-  const unsigned char *BinaryStart;
-  /// Pointer to the target code end
-  const unsigned char *BinaryEnd;
-  /// the offload entry table
-  _pi_offload_entry EntriesBegin;
-  _pi_offload_entry EntriesEnd;
-  // Array of preperty sets; e.g. specialization constants symbol-int ID map is
-  // propagated to runtime with this mechanism.
-  pi_device_binary_property_set PropertySetsBegin;
-  pi_device_binary_property_set PropertySetsEnd;
-  // TODO Other fields like entries, link options can be propagated using
-  // the property set infrastructure. This will improve binary compatibility and
-  // add flexibility.
-};
-using pi_device_binary = pi_device_binary_struct *;
-
 // pi_buffer_region structure repeats cl_buffer_region, used for sub buffers.
 struct pi_buffer_region_struct {
   size_t origin;
@@ -1135,27 +1025,6 @@ struct pi_image_region_struct {
   size_t depth;
 };
 using pi_image_region = pi_image_region_struct *;
-
-// Offload binaries descriptor version supported by this library.
-static const uint16_t PI_DEVICE_BINARIES_VERSION = 1;
-
-/// This struct is a record of all the device code that may be offloaded.
-/// It must match the __tgt_bin_desc structure generated by
-/// the clang-offload-wrapper tool when their Version field match.
-struct pi_device_binaries_struct {
-  /// version of this structure - for backward compatibility;
-  /// all modifications which change order/type/offsets of existing fields
-  /// should increment the version.
-  uint16_t Version;
-  /// Number of device binaries in this descriptor
-  uint16_t NumDeviceBinaries;
-  /// Device binaries data
-  pi_device_binary DeviceBinaries;
-  /// the offload entry table (not used, for compatibility with OpenMP)
-  _pi_offload_entry *HostEntriesBegin;
-  _pi_offload_entry *HostEntriesEnd;
-};
-using pi_device_binaries = pi_device_binaries_struct *;
 
 // This union encapsulates the two external handles we currently support.
 // When choosing the correct field from the union we need to look at the value
